+++ tests/ToastTest.php
@@ -142,4 +142,51 @@ HTML;
 
         $this->assertContainsWithoutLE("jQuery('#w0').toast();", $options);
     }
+
+    /**
+     * @see https://github.com/yiisoft/yii2-bootstrap5/issues/36
+     */
+    public function testWidgetNoInitialization()
+    {
+        Toast::$counter = 0;
+        ob_start();
+        $toast = Toast::begin([
+            'title' => 'Toast title',
+            'clientOptions' => false,
+            'titleOptions' => ['tag' => 'h5', 'style' => ['text-align' => 'left']]
+        ]);
+        echo 'test';
+        Toast::end();
+        $out = ob_get_clean();
+
+        $this->assertFalse($toast->clientOptions);
+        $this->assertArrayNotHasKey(View::POS_READY,Yii::$app->view->js);
+    }
+
+    /**
+     * 
+     * @see https://github.com/yiisoft/yii2-bootstrap5/issues/36
+     */
+    public function testWidgetInitializationTrue()
+    {
+        Toast::$counter = 0;
+        ob_start();
+        $toast = Toast::begin([
+            'title' => 'Toast title',
+            'clientOptions' => true,
+            'titleOptions' => ['tag' => 'h5', 'style' => ['text-align' => 'left']]
+        ]);
+        echo 'test';
+        Toast::end();
+        $out = ob_get_clean();
+
+        $this->assertTrue($toast->clientOptions);
+        $this->assertArrayHasKey(View::POS_READY,Yii::$app->view->js);
+        $js = Yii::$app->view->js[View::POS_READY];
+
+        $this->assertInternalType(IsType::TYPE_ARRAY, $js);
+        $options = array_shift($js);
+
+        $this->assertContainsWithoutLE("jQuery('#w0').toast(true);", $options);
+    }
 }